updated the header & login part
