+++ src/components/Header/Header.styles.js
@@ -1,55 +1,10 @@
-// TODO : It needs to be broken in several Style files for each component
+import { makeStyles } from '@material-ui/core/styles';
 
-import { alpha, makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles((theme) => ({
-  tool: {
+const headerStyle = makeStyles({
+  container: {
     display: 'flex',
     justifyContent: 'space-between'
-  },
-
-  search: {
-    position: 'relative',
-    borderRadius: theme.shape.borderRadius,
-    backgroundColor: alpha(theme.palette.common.white, 0.15),
-    '&:hover': {
-      backgroundColor: alpha(theme.palette.common.white, 0.25)
-    },
-    marginLeft: 5,
-    width: '80%',
-    [theme.breakpoints.up('sm')]: {
-      marginLeft: theme.spacing(1),
-      width: 'auto'
-    }
-  },
-  searchIcon: {
-    padding: theme.spacing(0, 2),
-    height: '100%',
-    position: 'absolute',
-    pointerEvents: 'none',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  inputRoot: {
-    color: 'inherit'
-  },
-  inputInput: {
-    padding: theme.spacing(1, 1, 1, 0),
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      width: '18ch',
-      '&:focus': {
-        width: '25ch'
-      }
-    }
-  },
-  logo: {
-    width: '4rem',
-    height: '3rem'
   }
-}));
+});
 
-export default useStyles;
+export default headerStyle;
