+++ src/components/Header/Logo.style.js
@@ -1,0 +1,10 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const logoStyle = makeStyles({
+  logo: {
+    width: '4rem',
+    height: '3rem'
+  }
+});
+
+export default logoStyle;
