+++ src/components/SocialMedia/SocialMedia.style.js
@@ -3,13 +3,11 @@ import { makeStyles } from '@material-ui/core/styles';
 const useStyles = makeStyles((theme) => ({
   socialIconStyle: {
     display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center'
+    justifyContent: 'space-around'
   },
   menuButton: {
     marginRight: theme.spacing(2),
     edge: 'start',
-    color: 'inherit',
     ariaLabel: 'menu'
   }
 }));
